requirements.txt
pandas
numpy
matplotlib
scikit-learn
tensorflow
seaborn
keras-tuner
pillow
flask
flask-cors
flasgger